requirements.txt
qq-botpy~=1.2.1